+++ TablePro/Models/Settings/SyncSettings.swift
@@ -13,8 +13,6 @@ struct SyncSettings: Codable, Equatable {
     var syncConnections: Bool
     var syncGroupsAndTags: Bool
     var syncSettings: Bool
-    var syncQueryHistory: Bool
-    var historySyncLimit: HistorySyncLimit
     var syncPasswords: Bool
     var syncSSHProfiles: Bool
 
@@ -23,8 +21,6 @@ struct SyncSettings: Codable, Equatable {
         syncConnections: Bool,
         syncGroupsAndTags: Bool,
         syncSettings: Bool,
-        syncQueryHistory: Bool,
-        historySyncLimit: HistorySyncLimit,
         syncPasswords: Bool = false,
         syncSSHProfiles: Bool = true
     ) {
@@ -32,8 +28,6 @@ struct SyncSettings: Codable, Equatable {
         self.syncConnections = syncConnections
         self.syncGroupsAndTags = syncGroupsAndTags
         self.syncSettings = syncSettings
-        self.syncQueryHistory = syncQueryHistory
-        self.historySyncLimit = historySyncLimit
         self.syncPasswords = syncPasswords
         self.syncSSHProfiles = syncSSHProfiles
     }
@@ -44,8 +38,6 @@ struct SyncSettings: Codable, Equatable {
         syncConnections = try container.decode(Bool.self, forKey: .syncConnections)
         syncGroupsAndTags = try container.decode(Bool.self, forKey: .syncGroupsAndTags)
         syncSettings = try container.decode(Bool.self, forKey: .syncSettings)
-        syncQueryHistory = try container.decode(Bool.self, forKey: .syncQueryHistory)
-        historySyncLimit = try container.decode(HistorySyncLimit.self, forKey: .historySyncLimit)
         syncPasswords = try container.decodeIfPresent(Bool.self, forKey: .syncPasswords) ?? false
         syncSSHProfiles = try container.decodeIfPresent(Bool.self, forKey: .syncSSHProfiles) ?? true
     }
@@ -55,35 +47,7 @@ struct SyncSettings: Codable, Equatable {
         syncConnections: true,
         syncGroupsAndTags: true,
         syncSettings: true,
-        syncQueryHistory: true,
-        historySyncLimit: .entries500,
         syncPasswords: false,
         syncSSHProfiles: true
     )
-}
-
-/// Maximum number of query history entries to sync
-enum HistorySyncLimit: String, Codable, CaseIterable {
-    case entries100 = "100"
-    case entries500 = "500"
-    case entries1000 = "1000"
-    case unlimited = "unlimited"
-
-    var displayName: String {
-        switch self {
-        case .entries100: return "100"
-        case .entries500: return "500"
-        case .entries1000: return "1,000"
-        case .unlimited: return String(localized: "Unlimited")
-        }
-    }
-
-    var limit: Int? {
-        switch self {
-        case .entries100: return 100
-        case .entries500: return 500
-        case .entries1000: return 1_000
-        case .unlimited: return nil
-        }
-    }
 }